+++ src/cli/commands/build.js
@@ -1,12 +1,12 @@
-import { runLegacyCLI } from '../run-legacy';
+import { callback, options, runLegacyCLI } from '../run-legacy';
 
 export default {
+	callback,
 	desc: 'Builds a project',
 	options: [
 		{
-			'-d, --project-dir [path]': 'The directory containing the project; defaults to the current directory',
-			'-f, --force':              'Force a full rebuild',
-			'-p, --platform [name]':    'The target build platform'
+			...options,
+			'-f, --force': 'Force a full rebuild'
 		}
 	],
 	async action(ctx) {