+++ src/cli/commands/run.js
@@ -1,13 +1,13 @@
-import { runLegacyCLI } from '../run-legacy';
+import { callback, options, runLegacyCLI } from '../run-legacy';
 
 export default {
+	callback,
 	desc: 'Build and runs a project',
 	options: [
 		{
-			'--build-only':             'Builds the project without running it in the simulator/emulator or installing it on device',
-			'-d, --project-dir [path]': 'The directory containing the project; defaults to the current directory',
-			'-f, --force':              'Force a full rebuild',
-			'-p, --platform [name]':    'The target build platform'
+			...options,
+			'--build-only': 'Builds the project without running it in the simulator/emulator or installing it on device',
+			'-f, --force': 'Force a full rebuild'
 		}
 	],
 	async action(ctx) {